backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.2

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Development
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
